requirements.txt
sentence-transformers
faiss-cpu
scikit-learn
numpy
pandas
fastapi
uvicorn
matplotlib
pydantic